combine batch_size_like.cc into batch_size_like.h
